ocr-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-multipart==0.0.6
pydantic==2.5.0 